Giving variants its own query module.
