+++ cibyl/sources/zuul/queries/jobs.py
@@ -41,15 +41,3 @@ def perform_jobs_query(zuul, **kwargs):
         result += jobs.get()
 
     return result
-
-
-def perform_variants_query(job, **kwargs):
-    """Query for variants.
-
-    :param job: API to interact with the owner of the variants.
-    :type job: :class:`cibyl.sources.zuul.transactions.JobResponse`
-    :param kwargs: See :func:`handle_query`.
-    :return: List of retrieved variants.
-    :rtype: list[:class:`cibyl.sources.zuul.transactions.VariantResponse`]
-    """
-    return job.variants().get()
